move catalog related services into a namespace; update tests; allow folioreader to run with a barcode lookup
