+++ .rubocop_todo.yml
@@ -1,6 +1,6 @@
 # This configuration was generated by
 # `rubocop --auto-gen-config`
-# on 2023-02-23 19:39:15 UTC using RuboCop version 1.45.1.
+# on 2023-02-24 21:47:51 UTC using RuboCop version 1.46.0.
 # The point is for the user to remove these configuration records
 # one by one as the offenses are removed from the code base.
 # Note that changes in the inspected code, or installation of new
@@ -12,7 +12,7 @@ Lint/SuppressedException:
   Exclude:
     - 'lib/tasks/rspec.rake'
 
-# Offense count: 66
+# Offense count: 67
 # Configuration parameters: AllowedMethods, AllowedPatterns, CountRepeatedAttributes.
 Metrics/AbcSize:
   Max: 39
@@ -22,12 +22,12 @@ Metrics/AbcSize:
 Metrics/ClassLength:
   Max: 114
 
-# Offense count: 12
+# Offense count: 13
 # Configuration parameters: AllowedMethods, AllowedPatterns.
 Metrics/CyclomaticComplexity:
   Max: 11
 
-# Offense count: 73
+# Offense count: 74
 # Configuration parameters: CountComments, CountAsOne, AllowedMethods, AllowedPatterns.
 Metrics/MethodLength:
   Max: 28
@@ -37,7 +37,7 @@ Metrics/MethodLength:
 Metrics/ParameterLists:
   Max: 6
 
-# Offense count: 9
+# Offense count: 10
 # Configuration parameters: AllowedMethods, AllowedPatterns.
 Metrics/PerceivedComplexity:
   Max: 12
@@ -45,7 +45,7 @@ Metrics/PerceivedComplexity:
 # Offense count: 11
 Naming/AccessorMethodName:
   Exclude:
-    - 'app/services/marc_856_generator.rb'
+    - 'app/services/catalog/marc_856_generator.rb'
 
 # Offense count: 9
 # Configuration parameters: EnforcedStyle, AllowedIdentifiers, AllowedPatterns.
@@ -53,6 +53,12 @@ Naming/AccessorMethodName:
 Naming/VariableName:
   Exclude:
     - 'app/services/publish/public_desc_metadata_service.rb'
+
+# Offense count: 1
+# Configuration parameters: MinSize.
+Performance/CollectionLiteralInLoop:
+  Exclude:
+    - 'app/services/catalog/symphony_reader.rb'
 
 # Offense count: 17
 RSpec/AnyInstance:
@@ -130,14 +136,14 @@ RSpec/StubbedMock:
 # Offense count: 2
 RSpec/SubjectStub:
   Exclude:
-    - 'spec/services/update_marc856_record_service_spec.rb'
+    - 'spec/services/catalog/update_marc856_record_service_spec.rb'
 
 # Offense count: 1
 # Configuration parameters: EnforcedStyle.
 # SupportedStyles: slashes, arguments
 Rails/FilePath:
   Exclude:
-    - 'app/services/marc_service.rb'
+    - 'app/services/catalog/marc_service.rb'
 
 # Offense count: 5
 # This cop supports unsafe autocorrection (--autocorrect-all).
@@ -220,7 +226,7 @@ Style/SymbolArray:
     - 'config/routes.rb'
     - 'lib/tasks/rspec.rake'
 
-# Offense count: 430
+# Offense count: 444
 # This cop supports safe autocorrection (--autocorrect).
 # Configuration parameters: AllowHeredoc, AllowURI, URISchemes, IgnoreCopDirectives, AllowedPatterns.
 # URISchemes: http, https
