+++ app/controllers/marcxml_controller.rb
@@ -2,10 +2,10 @@
 
 class MarcxmlController < ApplicationController # :nodoc:
   def marcxml
-    render xml: MarcService.marcxml(barcode: params[:barcode], catkey: params[:catkey])
-  rescue MarcService::CatalogRecordNotFoundError => e
+    render xml: Catalog::MarcService.marcxml(barcode: params[:barcode], catkey: params[:catkey])
+  rescue Catalog::MarcService::CatalogRecordNotFoundError => e
     json_api_error(status: :bad_request, title: 'Record not found in Symphony', message: e.message)
-  rescue MarcService::MarcServiceError => e
+  rescue Catalog::MarcService::MarcServiceError => e
     json_api_error(status: :internal_server_error, message: e.message)
   end
 end
