+++ spec/services/catalog/update_marc856_record_service_spec.rb
@@ -2,7 +2,7 @@
 
 require 'rails_helper'
 
-RSpec.describe UpdateMarc856RecordService do
+RSpec.describe Catalog::UpdateMarc856RecordService do
   subject(:umrs) { described_class.new(cocina_object, thumbnail_service:) }
 
   let(:druid) { 'druid:bc123dg9393' }
@@ -23,8 +23,8 @@ RSpec.describe UpdateMarc856RecordService do
 
   describe '#update' do
     before do
-      allow(Marc856Generator).to receive(:create).and_return(marc_856_records)
-      allow(SymphonyWriter).to receive(:save)
+      allow(Catalog::Marc856Generator).to receive(:create).and_return(marc_856_records)
+      allow(Catalog::SymphonyWriter).to receive(:save)
       umrs.update
     end
 
@@ -32,7 +32,7 @@ RSpec.describe UpdateMarc856RecordService do
       let(:marc_856_records) { ['abcd1244', 'def5678'] }
 
       it 'calls symphony_writer' do
-        expect(SymphonyWriter).to have_received(:save).with(marc_856_records)
+        expect(Catalog::SymphonyWriter).to have_received(:save).with(marc_856_records)
       end
     end
 
@@ -40,7 +40,7 @@ RSpec.describe UpdateMarc856RecordService do
       let(:marc_856_records) { [] }
 
       it 'does not call symphony_writer' do
-        expect(SymphonyWriter).not_to have_received(:save)
+        expect(Catalog::SymphonyWriter).not_to have_received(:save)
       end
     end
   end
